Change debug logging conditional
Now switches based on the default loggers implementation - allowing users to bypass and override to enable debug logging; while maintaining the previous functionality of the logging build flag
